requirements.txt
Django==4.2.0
psycopg2==2.9.9
SQLAlchemy==2.0.23
python-dotenv==1.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
django-bootstrap-v5==1.0.11
tzdata==2023.3
cffi==1.16.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-simple-history==3.7.0